Update 003_project_identifier_to_id.rb
Optimize db script for migration.

+++ db/migrate/003_project_identifier_to_id.rb
@@ -1,46 +1,35 @@
 class ProjectIdentifierToId < ActiveRecord::Migration[4.2]
   def up
-    add_column :issue_todo_lists, :project_id, :integer, null: true, after: :id
+    # Add new column if it doesn't exist
+    unless column_exists?(:issue_todo_lists, :project_id)
+      add_column :issue_todo_lists, :project_id, :integer, null: true, after: :id
 
-    # project_identifier => project_id (as identifier is not unique use LIMIT)
-    IssueTodoList.all.each do |list|
-      list.update(project_id: Project.where(identifier: list.project_identifier).first&.id)
+      # Convert project_identifier to project_id
+      IssueTodoList.all.each do |list|
+        list.update(project_id: Project.where(identifier: list.project_identifier).first&.id)
+      end
     end
 
-    remove_column :issue_todo_lists, :project_identifier
+    # Remove old column if it exists
+    if column_exists?(:issue_todo_lists, :project_identifier)
+      remove_column :issue_todo_lists, :project_identifier
+    end
   end
 
   def down
-    add_column :issue_todo_lists, :project_identifier, :string, null: true, after: :id
+    # Revert back - Add old column if it doesn't exist
+    unless column_exists?(:issue_todo_lists, :project_identifier)
+      add_column :issue_todo_lists, :project_identifier, :string, null: true, after: :id
 
-    # project_id => project_identifier
-    IssueTodoList.all.each do |list|
-      list.update(project_identifier: Project.where(id: list.project_id).first&.identifier)
+      # Convert project_id back to project_identifier
+      IssueTodoList.all.each do |list|
+        list.update(project_identifier: Project.where(id: list.project_id).first&.identifier)
+      end
     end
 
-    remove_column :issue_todo_lists, :project_id
-  end
-end
-class ProjectIdentifierToId < ActiveRecord::Migration[4.2]
-  def up
-    add_column :issue_todo_lists, :project_id, :integer, null: true, after: :id
-
-    # project_identifier => project_id (as identifier is not unique use LIMIT)
-    IssueTodoList.all.each do |list|
-      list.update(project_id: Project.where(identifier: list.project_identifier).first&.id)
+    # Remove the new column if it exists
+    if column_exists?(:issue_todo_lists, :project_id)
+      remove_column :issue_todo_lists, :project_id
     end
-
-    remove_column :issue_todo_lists, :project_identifier
-  end
-
-  def down
-    add_column :issue_todo_lists, :project_identifier, :string, null: true, after: :id
-
-    # project_id => project_identifier
-    IssueTodoList.all.each do |list|
-      list.update(project_identifier: Project.where(id: list.project_id).first&.identifier)
-    end
-
-    remove_column :issue_todo_lists, :project_id
   end
 end
